DAI_tpzrp-main/DAI_tpzrp-main/auth/requirements.txt
bcrypt
cryptography
email-validator
databases==0.9.0
fastapi
passlib
psycopg2-binary
pydantic
pydantic_core
PyJWT
SQLAlchemy
python-dotenv
uvicorn
python-multipart
pika
requests